requirements.txt
# Core data science and analysis libraries
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Google APIs
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-analytics-data>=0.18.0
google-analytics-admin>=0.22.0
google-ads>=22.0.0
oauth2client>=4.1.3

# Database connectors
mysql-connector-python>=8.0.0

# Web framework (optional)
flask>=2.3.0
fastapi>=0.100.0

# Database (optional)
sqlalchemy>=2.0.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
python-dotenv>=1.0.0

# PDF generation
reportlab>=4.0.0

# Jupyter for interactive development
jupyter>=1.0.0
ipykernel>=6.0.0